Rename all Files properly
